Only set up GMSL link per max9295
Calling the GMSL APIs only on the first sensor based on the device tree
link to max9295, so for a typical 2 D457 on link A/B to the same
max9296, these initial setting will be executed twice like the reference
drivers.

This should be considered as a temporary solution. Merging 4 sensors
into one D457 device or changing GMSL API should be better.

Signed-off-by: Xin Zhang <xin.x.zhang@intel.com>
Signed-off-by: Junze Wu <junze.wu@intel.com>